Stable Kernel Execution API
